+++ src/main/kotlin/dev/mayaqq/estrogen/config/EstrogenServerConfig.kt
@@ -30,6 +30,9 @@ object EstrogenServerConfig {
 
         @Comment("Chance for a random tick to cause the dreaming effect to be applied in percent")
         var dreamingTickChance: @Range(0, 100) Int = 100
+
+        @Comment("Range from a dream catcher for a player to be to block the Dreaming Effect")
+        var dreamCatcherRange: Int = 10
     }
 
     @Comment("Settings which are more fun and not fit for survival")