+++ src/main/kotlin/dev/mayaqq/estrogen/content/Transgenders.kt
@@ -5,6 +5,7 @@ import dev.engine_room.flywheel.api.visualization.BlockEntityVisualizer
 import dev.engine_room.flywheel.api.visualization.VisualizationContext
 import dev.engine_room.flywheel.api.visualization.VisualizerRegistry
 import dev.engine_room.flywheel.lib.visualization.SimpleBlockEntityVisualizer
+import dev.mayaqq.cynosure.blocks.poi.PoiHelpers
 import dev.mayaqq.cynosure.items.extensions.CustomTooltip
 import dev.mayaqq.cynosure.items.extensions.registerExtension
 import dev.mayaqq.cynosure.tooltips.DescriptionTooltip
@@ -14,11 +15,12 @@ import earth.terrarium.baubly.client.BaubleRenderer
 import earth.terrarium.baubly.client.BaublyClient
 import earth.terrarium.baubly.common.Bauble
 import net.minecraft.client.renderer.item.ClampedItemPropertyFunction
-import net.minecraft.client.renderer.item.ItemProperties
 import net.minecraft.resources.ResourceLocation
+import net.minecraft.world.entity.ai.village.poi.PoiType
 import net.minecraft.world.item.Item
 import net.minecraft.world.level.block.entity.BlockEntity
 import net.minecraft.world.level.block.entity.BlockEntityType
+import uwu.serenity.kritter.api.builders.GenericBuilder
 import uwu.serenity.kritter.client.stdlib.clientOnly
 import uwu.serenity.kritter.stdlib.BlockEntityBuilder
 import uwu.serenity.kritter.stdlib.ItemBuilder
@@ -75,6 +77,13 @@ inline fun <BE : BlockEntity> BlockEntityBuilder<BE>.visualizaer(crossinline fac
         onSetup {
             VisualizerRegistry.setVisualizer(it, factory(it))
         }
+    }
+}
+
+// Pois
+fun GenericBuilder<PoiType, PoiType>.registerState() {
+    onRegister {
+        PoiHelpers.registerState(it)
     }
 }