+++ src/main/kotlin/dev/mayaqq/estrogen/content/recipes/CreativelessShapedRecipeClient.kt
@@ -1,7 +1,0 @@
-package dev.mayaqq.estrogen.content.recipes
-
-import net.minecraft.client.Minecraft
-
-object CreativelessShapedRecipeClient {
-    fun isInCreative(): Boolean = Minecraft.getInstance().player?.isCreative == true
-}