+++ CHANGELOG.md
@@ -5,3 +5,5 @@
 - Dream Block Rework (Sery)
 - Moth Elytra Flapping
 - Doubled Default Time it takes for patches to Drain
+- Dream Catcher
+- Moth Wool Carpets renamed to just Moth Carpets